Move Difference and ListDifference to json package, convert them to value classes
